--- Namespace/app.ts
@@ -1,4 +1,0 @@
-// using refer.ts file 
-/// <reference path = "refer.ts" />
-
-console.log(Book.getName()," is a book written by ",Writer.getName());  // calling functions from namespace file - refer.ts
