--- Namespace/readme.md.txt
@@ -1,5 +1,0 @@
-Namespaces are paradigm of organizing code so that variables,functions, interfaces, or classes are grouped together within a local scope in order to avoid naming 
-conflicts between components in the global scope.
-
-To run the following - 
-node ./out.js
